wrote documentation for changes to pathgenerator and fixed or moved code based on prefix change
