--- mirrulations_pathgenerator/pathgenerator.py
@@ -1,189 +1,0 @@
-class PathGenerator:
-    """
-    A Class which classifies any type of file into the correct directory
-    following our data structure where every path is based on an agencyId, id,
-    docketId and type.
-    Paths are generated by extracting information from a dockets, documents or
-    comments .json file
-
-    Important Keys:
-    ----------
-    agencyId: The Governmental Agency Abbreviation from Regulations.gov
-
-    id: Everything related to a docket is identified by its id
-        Ex of a Docket's 'id': USTR-2015-0010
-        Ex of a Document's 'id':  USTR-2015-0010-0001
-        Ex of a Comment's 'id': USTR-2015-0010-0002
-    type: The value for a json's 'type' key
-        Can be a docket, document, comment or attachment
-    ...
-    Methods
-    -------
-    get_path(json = dict)
-        Looks for the 'type' key in the json and calls the appropriate path
-        generator method for the given 'type'
-    parse_docket_id(item_id = str)
-        If the json does not have the docketId field we must parse out the
-        docketId using split methods.
-    get_attributes(json_data, is_docket_json=False)
-        Returns the agencyid, docket_id, and item_id from a loaded json object.
-        Also handles if a key is not found in the json
-    get_docket_json_path(json = dict):
-        Gets the path for a json with the 'dockets' type
-    get_document_json_path(json = dict):
-        Gets the path for a json with the 'documents' type
-    get_comment_json_path(json = dict):
-        Gets the path for a json with the 'comments' type
-    """
-
-    def get_json_path(self, json):
-        if json['data']["type"] == "comments":
-            return self.get_comment_json_path(json)
-        if json['data']["type"] == "dockets":
-            return self.get_docket_json_path(json)
-        if json['data']["type"] == "documents":
-            return self.get_document_json_path(json)
-        return "/unknown/unknown.json"
-
-    def get_path(self, json):
-        if 'data' not in json or json['data'] == []:
-            return "unknown/unknown.json"
-        if json['data'].get("type") != -1:
-            return "Raw_data" + self.get_json_path(json)
-        return "unknown/unknown.json"
-
-    def _get_nested_keys_in_json(self, json_data, nested_keys, default_value):
-        '''
-        Gets a value from traversing a series of nested keys in a JSON object.
-        default_value is the value that should be returned if any of the nested
-        keys are missing from the JSON.
-        '''
-        json_subset = json_data
-
-        for key in nested_keys:
-            if key not in json_subset:
-                return default_value
-            json_subset = json_subset[key]
-        return json_subset
-
-    def parse_docket_id(self, item_i_d):
-        if item_i_d is None:
-            return "unknown"
-
-        segments = item_i_d.split('-')  # list of segments separated by '-'
-        segments_excluding_end = segments[:-1]  # drops the last segment
-        parsed_docket_id = '-'.join(segments_excluding_end)
-        return parsed_docket_id
-
-    def _check_for_none_values(self, docket_id, agency_id, ):
-        if docket_id is None:
-            docket_id = 'unknown'
-        if agency_id is None:
-            agency_id = 'unknown'
-        return docket_id, agency_id
-
-    def get_attributes(self, json_data, is_docket=False):
-        '''
-        Returns the agency, docket id, and item id from a loaded json object.
-        '''
-        item_i_d = self._get_nested_keys_in_json(
-            json_data, ['data', 'id'], None)
-        agency_id = self._get_nested_keys_in_json(
-            json_data, ['data', 'attributes', 'agencyId'], None)
-
-        if is_docket:
-            docket_id = item_i_d
-            item_i_d = None
-        else:
-            docket_id = self._get_nested_keys_in_json(
-                json_data, ['data', 'attributes', 'docketId'], None)
-
-            if docket_id is None:
-                docket_id = self.parse_docket_id(item_i_d)
-        if not is_docket and item_i_d is None:
-            item_i_d = 'unknown'
-        docket_id, agency_id = self._check_for_none_values(docket_id,
-                                                           agency_id)
-
-        return agency_id, docket_id, item_i_d
-
-    def get_docket_json_path(self, json):
-        agency_i_d, docket_i_d, _ = self.get_attributes(json,
-                                                        is_docket=True)
-        return f'/{agency_i_d}/{docket_i_d}/text-{docket_i_d}/docket/' + \
-               f'{docket_i_d}.json'
-
-    def get_document_json_path(self, json):
-        agency_i_d, docket_i_d, item_i_d = self.get_attributes(json)
-
-        return f'/{agency_i_d}/{docket_i_d}/text-{docket_i_d}/documents' + \
-               f'/{item_i_d}.json'
-
-    def get_document_htm_path(self, json):
-        agency_id, docket_id, item_id = self.get_attributes(json)
-
-        return f'/{agency_id}/{docket_id}/text-{docket_id}/' + \
-               f'documents/{item_id}_content.htm'
-
-    def get_comment_json_path(self, json):
-        agency_i_d, docket_i_d, item_i_d = self.get_attributes(json)
-
-        return f'/{agency_i_d}/{docket_i_d}/text-{docket_i_d}/comments/' + \
-               f'{item_i_d}.json'
-
-    def _has_file_formats(self, attributes, attachment):
-        if attributes.get("fileFormats"):
-            return True
-        print("fileFormats did not exist for attachment ID: " +
-              f"{attachment.get('id')}")
-        return False
-
-    def _parse_attachment_path(self, json, file_format, attachments):
-        agency_i_d, docket_i_d, item_i_d = self.get_attributes(json)
-        if "fileUrl" in file_format:
-            attachment_name = item_i_d + "_" + \
-                file_format["fileUrl"].split("/")[-1]
-            attachments.append(f'/{agency_i_d}/{docket_i_d}/' +
-                               f'binary-{docket_i_d}/comments_' +
-                               f'attachments/{attachment_name}')
-        return attachments
-
-    def get_attachment_json_paths(self, json):
-        '''
-        Given a json, this function will return all attachment paths for
-        n number attachment links
-        '''
-
-        # contains list of paths for attachments
-        attachments = []
-        for attachment in json["included"]:
-            attributes = attachment["attributes"]
-            if self._has_file_formats(attributes, attachment):
-                for file_format in attributes["fileFormats"]:
-                    attachments = self._parse_attachment_path(json,
-                                                              file_format,
-                                                              attachments)
-
-        return attachments
-
-    @staticmethod
-    def make_attachment_save_path(path):
-        '''
-        This method takes a complete path to a pdf and makes
-        the save path based on parameters in that path.
-        Parameters
-        ----------
-        path : str
-            the complete file path for the attachment that is being extracted
-            ex. /path/to/pdf/attachment_1.pdf
-        '''
-        if "comments" in path:
-            return path.replace('binary', 'text') \
-                       .replace('comments_attachments',
-                                'comments_extracted_text/pdfminer') \
-                       .replace('.pdf', '_extracted.txt') \
-
-        return path.replace('binary', 'text') \
-                   .replace('documents_attachments',
-                            'documents_extracted_text/pdfminer') \
-                   .replace('.pdf', '_extracted.txt')
